useInitializationList
When an object of a class is created, the constructors of all member
variables are called consecutively in the order the variables are
declared, even if you don't explicitly write them to the
initialization list. You could avoid assigning 'myItem' a value by
passing the value to the constructor in the initialization list.